Refactor alexa modules to avoid circula deps
